# apps/backend/src/market/routes.ts
 
import { Router } from 'express';
import { exchangeService } from '../exchanges/ExchangeService';

export const marketRouter = Router();

// Поиск символов
marketRouter.get('/search', async (req, res) => {
  const { exchange, query } = req.query;
  try {
    const adapter = exchangeService.getAdapter(exchange as string);
    const symbols = await adapter.searchSymbols(query as string);
    res.json({ symbols });
  } catch (e: any) {
    res.status(400).json({ error: e.message });
  }
});

// Исторические свечи (для старта графика и пагинации)
marketRouter.get('/history', async (req, res) => {
  const { exchange, symbol, tf, from, limit } = req.query;
  
  try {
    const adapter = exchangeService.getAdapter(exchange as string);
    const candles = await adapter.getHistoricalCandles(
      symbol as string, 
      tf as any, 
      from ? Number(from) : undefined,
      limit ? Number(limit) : 500
    );
    res.json({ candles });
  } catch (e: any) {
    res.status(400).json({ error: e.message });
  }
});